+++ App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
-import SearchResults from './src/components/SearchResults';
-import ProductInfo from './src/components/ProductInfo';
-import RootNav from './src/components/navigator.js';
+import SearchScreen from './src/components/SearchScreen.js';
+import ProductInfo from './src/components/ProductInfo.js';
+import RootNav from './src/components/Navigator.js';
 
 
 export default class App extends React.Component {
