+++ src/components/AllergyScreen.js
