+++ src/components/searchscreen.js
@@ -14,39 +14,76 @@ function adjusted(names) {
 class SearchScreen extends Component {
   constructor() {
     super();
-    
     this.state = {
       selected: '',
-      items: [
-        'Goa',
-        'Gujrat',
-        'Madhya Pradesh',
-        'Assam',
-        'Gujrat',
-        'Karnataka',
-        'Jharkhand',
-        'Himachal Pradeshaaaaaaaaaaaaaaaaaaaaaaaakkkkkkkkkkkkkkkkkkkkkkkkkkkkkkk',
-        'Manipur',
-        'Meghalaya',
-        'Mizoram',
-        'Uttarakhand',
-        'West Bengal',
-        'Tamil Nadu ',
-        'Punjab',
-        'Rajasthan',
-        'Bihar',
-        'Andhra Pradesh',
-        'Arunachal Pradesh',
-        'a',
-        'b',
-        'c',
-        'd',
-        'e',
-        'f',
-        'g',
-        'h',
-      ]
+      items: []
     };
+  }
+
+
+  componentDidMount() {
+    var searchTerm = this.props.navigation.state.params.data.searchTerm;
+    console.log(searchTerm);
+
+    // Construct URL for database query from user's search entry
+    var splitSearch = searchTerm.split(" ");
+    var urlPart1 = "https://ndb.nal.usda.gov/ndb/search/list?fgcd=&manu=&lfacet=&count=&max=50&sort=default&qlookup=";
+    var urlPart2 = "&offset=&format=Full&new=&measureby=&ds=&order=asc&qt=&qp=&qa=&qn=&q=&ing="
+    var urlSearch = "";
+    for (var i = 0; i < splitSearch.length-1; i++) {
+      urlSearch += splitSearch[i] + "+";
+    }
+    urlSearch += splitSearch[splitSearch.length-1];
+    var url = urlPart1 + urlSearch + urlPart2;
+
+    fetch(url)
+       .then(res => {
+          if (res.ok) {
+            var str = JSON.stringify(res);
+            var arr = str.split('reports for this food');
+            var names = [];
+            var ndbs = [];
+            var final = [];
+
+            // Populate names array with names
+            for (var i = 0; i < arr.length; i += 2) {
+              var j = arr[i].indexOf("<");
+              element = arr[i].substring(10, j-10);
+              trimmed = element.trim();
+
+              // NOTE: special characters (&amp) aren't formatted
+
+              names.push(trimmed);
+            }
+
+            names.shift(); // Remove garbage element
+
+            // Populate ndbs array with NDBs
+            for (var i = 1; i < arr.length; i += 2) {
+              var j = arr[i].indexOf("<");
+              element = arr[i].substring(10, j);
+              trimmed = element.trim();
+              ndbs.push(trimmed);
+            }
+            
+            // Populate final with combined name/ndb objects
+            for (var i = 0; i < names.length; i++) {
+              final.push({"name": names[i], "ndb": ndbs[i]})
+            }
+
+            console.log("FINAL ARRAY: ");
+            for (var i = 0; i < final.length; i++) {
+              console.log(final[i]);
+            }
+
+            this.setState({items: final});
+            return final;
+
+          }
+        else {
+          throw new Error(res)
+        }
+      })
   }
 
   //<Text onPress={this.handleSelectedPress}
@@ -61,9 +98,9 @@ class SearchScreen extends Component {
             <Button 
               onPress={ 
                 () => {this.setState({selected: {item}})
-                navigate("Home")}
+                navigate("ProductInfo", {data: {ndb: item.ndb}})}
               }
-              title = {item}
+              title = {item.name}
               className = "text-left"/>
           </View>
         ))}
